requirements.txt
aiogram>=3.0.0
apscheduler
pytz
python-dotenv